Progress is now float

+++ src/Manisero.Navvy/BasicProcessing/BasicTaskStep.cs
@@ -10,7 +10,7 @@ namespace Manisero.Navvy.BasicProcessing
         /// <inheritdoc />
         public Func<TaskOutcome, bool> ExecutionCondition { get; }
 
-        public Action<IProgress<byte>, CancellationToken> Body { get; }
+        public Action<IProgress<float>, CancellationToken> Body { get; }
 
         /// <param name="executionCondition">See <see cref="ExecutionCondition"/>. If null, <see cref="TaskStepUtils.DefaultExecutionCondition"/> will be used.</param>
         public BasicTaskStep(
@@ -21,10 +21,11 @@ namespace Manisero.Navvy.BasicProcessing
         {
         }
 
+        /// <param name="body">Reported progress values should be between 0.0f and 1.0f (1.0f meaning 100%).</param>
         /// <param name="executionCondition">See <see cref="ExecutionCondition"/>. If null, <see cref="TaskStepUtils.DefaultExecutionCondition"/> will be used.</param>
         public BasicTaskStep(
             string name,
-            Action<IProgress<byte>, CancellationToken> body,
+            Action<IProgress<float>, CancellationToken> body,
             Func<TaskOutcome, bool> executionCondition = null)
         {
             Name = name;
@@ -50,11 +51,12 @@ namespace Manisero.Navvy.BasicProcessing
             => new BasicTaskStep(name, body, executionCondition);
 
         /// <summary>Builds <see cref="BasicTaskStep"/>.</summary>
+        /// <param name="body">Reported progress values should be between 0.0f and 1.0f (1.0f meaning 100%).</param>
         /// <param name="executionCondition">See <see cref="ITaskStep.ExecutionCondition"/>. If null, <see cref="TaskStepUtils.DefaultExecutionCondition"/> will be used.</param>
         public static BasicTaskStep Basic(
             this TaskStepBuilder _,
             string name,
-            Action<IProgress<byte>, CancellationToken> body,
+            Action<IProgress<float>, CancellationToken> body,
             Func<TaskOutcome, bool> executionCondition = null)
             => new BasicTaskStep(name, body, executionCondition);
     }
